move:#5 파일 이동(supabase client, 데이터 가져오는 로직)
