+++ src/app/api/supabase/service.ts
@@ -1,5 +1,6 @@
 import { ContactItemType, ContactWithPlansDetailType } from '@/types/contacts';
 import { supabase } from '@/app/api/supabase/client';
+import { PlansType } from '@/types/plans';
 
 // contacts 데이터 가져오기
 export const getContacts = async (userId: string): Promise<ContactItemType[]> => {
@@ -42,4 +43,15 @@ export const getContactsWithPlans = async (userId: string, contactsId: string): 
   }
 
   return { contact: data, plans: data.plans || [] };
+};
+
+// plans 데이터 가져오기 - calendar 사용
+export const getPlans = async (): Promise<PlansType[]> => {
+  const { data: plans, error } = await supabase
+    .from('plans')
+    .select('plan_id, user_id, contacts_id, title, detail, priority, start_date, end_date');
+  if (error) {
+    throw new Error(error.message);
+  }
+  return plans;
 };
