--- src/lib/utils/supabaseApi.ts
@@ -1,10 +1,0 @@
-import { PlansType } from '@/types/plans';
-import { supabase } from '@/lib/utils/supabaseClient';
-
-export const getPlans = async (): Promise<PlansType[]> => {
-  const { data: plans, error } = await supabase.from('plans').select('*');
-  if (error) {
-    throw new Error(error.message);
-  }
-  return plans;
-};
